Account database and methods to modify
Created a db for storing user variables. Added methods in AccountDb.java for:
Adding new accounts, authenticating a user, accessing and modifying scores in respective categories, updating the leaderboard for each category, adding and removing friends, listing user data

I made name a unique variable, so no 2 users can have the same name, because users need to enter something unique for the program to access their account and not someone else's, and giving their IDs to them at the time of registration is just bad. the main method is just for testing the code, it doesnt do anything else